internal: podlist: wait: fix ForPodList semantics
in `internal/wait` we have a couple functions with inconsistent
signature:
- ForPodListAllRunning
- ForPodListAllDeleted

the former takes and return slices of pointers to `v1.Pod`; the latter
takes a slice of `v1.Pod`. This was likely an implementation oversight.

The problem however run deeper: what's the correct signature?
Let's take a step back.
`internal/podlist` builds upon `sigs.k8s.io`'s controller-runtime,
which uses` PodList to hold `Items`, which is a slice of `v1.Pod`.
There is no reason to differ here, so for us a `podlist` should be
akin to a slice of `v1.Pod`.

Thus, functions which operate on a slice of pointers to `v1.Pod` should
use a different name, like maybe `Pods`.

Going back to the original question, and considering all the above,
the most correct API seems to be the one using pointers.
`ForPodListAllDeleted` should thus be updated to ensure the return
values of `ForPodListAllRunning` can be used directly.

Finally, both functions should be renamed to move from PodList to Pods.

Signed-off-by: Francesco Romani <fromani@redhat.com>