objectstate: use existing object as a base
Usually we're using a bare manifest as the base template for all the resources we own.
The operator performs the relevant modifications on those manifests and apply them on the cluster.

That usually creates unnecessary updates because the bare manifests are not
containing some of the defaults which applied by the API or other controllers
that present on the cluster.

To fix that, we suggest to use the existing resources as the base,
so those already contains all the defaults, perform the changes we care
about and then apply those on the cluster.

This is a different approach to accomplish:
https://github.com/openshift-kni/numaresources-operator/pull/1194

Signed-off-by: Talor Itzhak <titzhak@redhat.com>