lint: add gci linter to enforce import ordering
add and use the gci linter to enforce the import ordering.
The adopted ordering is not the nicest nor the best, but it's
the closest to our existing unwritten standard and can be
now enforced automatically, which is more than good enough.

Subtrees intentionally excluded by linting:
- api: because autogeneration
- cmd: because kubebuilder scaffolding requirement

GCI is: https://github.com/daixiang0/gci

Signed-off-by: Francesco Romani <fromani@redhat.com>