e2e: sched: assert expected scheduler pods amount
It isn't enough to check that the pods list of the scheduler is not
empty but rather we need to check that it has one pod exactly (assuming
single scheduler replica). The old assumption that the list would
contain one pod is misleading especially when we want to ensure that the
pod got recretaed and compare old UID with new UID.
Adjust teh assertion to verify that the list includes only one pod.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>