controller: sched: use current resource status as base for updated
syncNUMASchedulerResources() is constructing the sched status and later
in the reconcile loop this status is overriding the main instance status.
Concerning status conditions, the only situation where the
conditions are not set is on first creation of the scheduler CR,
the rest of the cases the conditions are there but they are continuously
overridden with nil in syncNUMASchedulerResources(). This in turn leads
to updating the conditions in r.updateStatus() whenever it is invoked
regardless of the actual existing conditions.

In this commit, ensure that the original current status is the base of
the new one.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit f3893f3de7c1ac2ac7f655400ed1e3f619369f71)
(cherry picked from commit ce4348690910e98488a65fc49cbe8b9c29e81fbe)
(cherry picked from commit 6ee6d4dcc52b1b8078e03a6b9826d58930bd0d58)