controller: sched: preserve current status.conditions
syncNUMASchedulerResources() is constructing the sched status and later
in the reconcile loop this status is overriding the main instance status.
Concerning status conditions, the only situation where the
conditions are not set (=nil) is on first creation of the scheduler CR,
the rest of the cases the conditions are there but they are continuously
overridden with nil in syncNUMASchedulerResources(). This in turn leads
to updating the conditions in r.updateStatus() whenever it is invoked
regardless of the actual existing conditions.

In this commit, ensure that the original current status is the base of
the new one.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit 7f755ba1ae09e9196b8fe13f8dc0b61a23099f95)