e2e:config: support old rte config format
rte configmap API has changed at 4.18
legacy version:
https://github.com/openshift-kni/numaresources-operator/blob/release-4.16/rte/pkg/config/config.go\#L41a
updated version (4.18+):
https://github.com/openshift-kni/numaresources-operator/blob/release-4.18/rte/pkg/config/config.go#L48

The test code is using the API structures from main (which is same as 4.18)
but the same test code is being exercised on versions which are older than 4.18.
The yaml.Umarshal call failed to parse the structure properly and that's why we're getting an empty string and the comparison failed.

In order to fix that we're adding a fallback method that will try to craft rte config from the
legacy config.

In addition replace Unmarshal -> UnmarshalStrict to catch the error
when data is the old format but the struct is the new format.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>