controller: sched: report DedicatedInformerActive behavior
We want to be able to see the informer mode of the scheduler as a
preparation step for having the informer default change according to the
kubelet version.
This commit enables reflecting the informer value (default or
configured) as part of the status conditions which is less direct than
reporting it in spec and is valid.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit 4d49a0a02c6bf92e6f5028e96fd2a2f2207854eb)