e2e: fixture: add a Log and Dump instances
Add a Log member in a fixture. It will use GinkgoLogr,
which is a go-logr compatible object which writes to
GinkgoWriter, which enables our desired integration
of debug messages with ginkgo.

Add an additional Dump interface with proper support for
multi-line output, which is a poor fit for structured logging,
but which is very helpful to troubleshoot the serial suite.

Signed-off-by: Francesco Romani <fromani@redhat.com>