test: internal: return non-pointer scheduler variable
CheckNROSchedulerAvailable() was returning a pointer to a scheduler
type and after it is invoked we normally would check that the result
isn't nil to tell if the functionality of the function passed or not.
The problem is that the returned value would never be nil. The only 2
options it would returns as a pointer are: 1. pointer to an empty
scheduler; 2. pointer to non-empty scheduler (the resources object from
the cluster).

This commit intend to remove the confusion and misuse of the function
and make it reurn a scheduler type instead of a pointer to the type. The
verififation of the result afterwards is by comparing it against empty
scheduler, and does the required adaptations accordingly.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>