scheduler: enable to consume the burstable setting
The current resource requirements set for the scheduler
pods were set tentatively long ago as initial estimation.

There are existing and acknowledged scalability concerns
tied to these settings. We are in the process of reviewing
the settings, and while the direction is pretty much clear,
the values are still being adjusted and reviewed.

In order to enable gradual rollout, we guard the new
settings behind a custom NUMAResourcesScheduler annotation,
which will act as pseudo-feature gate.

Once the new recommended values are solidified, we will
switch the behavior of the annotations, allowing to revert
to the old settings iff needed. Finally, mimicking
the real feature gate flow, we will remove the support
of the annotation.

Signed-off-by: Francesco Romani <fromani@redhat.com>