test: nrosched: clarify docstring for `CheckNROSchedulerAvailable()`
The function either succeeds and returns a pointer to the found
NUMA-aware scheduler, or the calling test fails. The function returned
value was asserted in some places as an attempt to ensure the functionality passed,
while that is inaccurate considering that the function already verifies its
self by being marked as ginkgo helper.
Clarify the intent of the function and remove redundant assertions.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>