controller: sched: update status only if needed
So far updating the scheduler status was done in every reconcilation iteration
even if the spec fields reflected the same values.

This commit saves a copy of the initial resource status and uses it as
a reference to identify if the resource requires a status update or not
and act accordingly.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit 558833b6ec807809f019a3529624780a014aa45d)