controller: sched: report DedicatedInformerActive behavior
We want to be able to see the informer mode of the scheduler as a
preparation step for having the informer default change according to the
kubelet version.
This commit enables reflecting the informer value (default or
configured) as part of the status conditions which is less direct than
reporting it in spec and is valid.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit 09d76b94c99d3eeec58863813570d0d377714aeb)
(cherry picked from commit 79be8cb0e49ce4612eb82edb2a10c7ed72d77c03)