controller: add PodAntiAffinity to RTE DS
We want to avoid having more than 1 RTE pod deployed on node when there
are overlapping nodeGroups specified in the NRO CR.

We cannot prevent the misconfiguration of the cluster node labels with
respect to MCO (e.i a node may match to more than one MCP's node
selector field) but we can mitigate having duplicated RTEs on the same
node.As a result, the order of RTE pods being created following
configuration modifications, like nodes labels, MCP node selectors or any
other adjustments that may relate to the specified nodeGroup(s), will
keep the pod pending starting the second one.

Important note: this change doesn't prevent running the RTE pod on nodes that do not
belong to the specified MCP but still match their nodeSelectors. The
only way to manage that is by not having overlapping node labels in the
first place. The operator depends on the MCO to group nodes of the
cluster into one configuration, but it should not duplicate how MCO
computes which node belongs to which MCP.

ref: https://kubernetes.io/docs/concepts/scheduling-eviction/assign-pod-node/#more-practical-use-cases

Signed-off-by: Shereen Haj <shajmakh@redhat.com>