controller: sched: adapt that kubelet PodResourcesAPI fix has landed
The fix has landed in OCP nightly 4.20, make sure this is reflected to
consume the new default of schedulerInformer in case it is not set by
the user.

Previously the version parsing relied on the assumption that the
platform version is of type stable (e.i 4.20.1, 4.20.2 ..) but this is
in correct because the cluster may be deployed with different types of
builds and each requires its own processing logic, which must know the
first supported build for each type. Ensure that the different types are
recognized and processed as required.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>