rbac review: move rte RBACs into rte
Instead of conflating all the RBAC annotations in the controller,
which was a historical artifact, we should either centralize them,
or do the opposite and split them and move close to the user.
The easiest path is the second one as the first one requires a deeper
redesign, so this PR implements it.

Signed-off-by: Francesco Romani <fromani@redhat.com>