sched: report computed replicas in status
the scheduler can compute automatically the scheduler
replicas (HA by default). It is better and friendlier
(for humans and programs) to have direct visibility
of the computed values in the NUMAResourcesScheduler status.

In general, all the autodetermined values should have
explicit feedback in the status, because this represent
the desired state.

Signed-off-by: Francesco Romani <fromani@redhat.com>