sched: status: reset the status conditions to defaults
The problem with the current implementation is that if the status condition A is
updated to `True` (like Progressing or Degraded condition) it will never
be updated again if the reconcilation interation is not updating that
exact condition. This creates accumulated status conditions as follows:

```
    conditions:
    - lastTransitionTime: "2025-09-09T10:24:02Z"
      message: ""
      reason: AsExpected
      status: "True"
      type: Available
    - lastTransitionTime: "2025-09-09T10:24:02Z"
      message: ""
      reason: Upgradeable
      status: "True"
      type: Upgradeable
    - lastTransitionTime: "2025-09-09T10:23:47Z"
      message: ""
      reason: AsExpected
      status: "True"
      type: Progressing
    - lastTransitionTime: "2025-09-09T10:23:47Z"
      message: 'could not apply (apps/v1, Kind=Deployment) openshift-numaresources/secondary-scheduler:
        could not update object (apps/v1, Kind=Deployment) openshift-numaresources/secondary-scheduler:
        Operation cannot be fulfilled on deployments.apps "secondary-scheduler": the
        object has been modified; please apply your changes to the latest version
        and try again'
      reason: InternalError
      status: "True"
      type: Degraded
    - lastTransitionTime: "2025-09-09T10:24:02Z"
      message: ""
      observedGeneration: 1
      reason: DedicatedInformerActive
      status: "False"
      type: DedicatedInformerActive
```

From which the user cannot know what condition is the true one.
To fix this, in the phase where the scheduler status is being updated,
make the status starts with default conditions instead of the old
ones to ensure a reset of the old conditions and a fresh start.
Also add e2e test coverage to existing test.

Note that a controller test cannot be added to verify the conditions
updates scenario because:
1.The only conditions that the instance status gets degraded is due to
K8s client returned errors, or invalid name, both cannot be used in this
scenario
2.scheduler deployment will never be ready (fake client) which leads to
   ->
3. scheduler condition is always Progressing because it is waiting for
   the deployment to be ready, which never happens.
That means there is no combination of two different condition statuses
that we can use to acheive this verification.

This is a reporting issue and doesn't affect the scheduler
functionality.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>
(cherry picked from commit f059c382772246ba36896b28da3217f2d48a0e07)