tools: introduce `pfpsyncchk`
Introduce new tool `pfpsyncchk` check that the PFP status dump provided by
RTE and scheduler are in sync, if not the binary prints out the
difference in the list of pods from both ends.
The tool uses the `record` package which is used originally on RTE and
secondary scheduler to store the captured PFP status.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>