e2e: serial: ensure MCPs and nodes are healthy
At the setup of the test, check that MCPs and nodes are healthy enough
for the test to run by verifying that the MCPs are all `Updated` and non of the nodes is
`unschedulable`. We don't want to fail on any error so keep any
unexpected results at warning level.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>