e2e:serial: fix baseload addition to pod resources
The old method used to apply the baseload on the same `paddingRes` over
and over again. This is not causing issues as long as there are no more
than one node to be padded, which is almost always the case on the MNO
used in the weekly CI.
However this is still incorrect, so this commit fixes it so that the
padding resources with the baseload are cleanly computed for each node.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>