e2e:serial: apply the baseload on the first numa zone
with the old version, the baseload considered in the padding resources
of every numazone. that is incorrect because it can create issues on
tests that expect their testing pods to be scheduled on one node only
while other nodes are unsuitable due to insufficient resources.
This update makes sure that only the forst NUMA zone s padded leaving
extra space for the baseload.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>