e2e:serial: pull daemonset object in eventually to ensure updates
Sometimes it takes longer time to have the daemonset updates its pods
amounts. The problem with the old code is that in an Eventually it waits
for the daemonset pods number to get updated to the expected number
while using an old daemonset object. To fix this, just pull the
daemonset object in every iteration to ensure the code in the eventaully
is using the most updated data.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>