e2e: serial: improve HaveLen failure output
We use HaveLen() to check that running and failed pods
matches expectation. Problem is that on failure the full
object output is too big and this exhaust the output buffer.
This means the most important information in the context
(status) is lost.

To fix this, we add a custom describe function which
focuses on the pod identification and the status.

Signed-off-by: Francesco Romani <fromani@redhat.com>