e2e: serial: fix panics
fix panic due to incorrect initialization,
by simplifying the code again.

The initialization of both `singleNUMAContainerScopeFuncs`
and `singleNUMAPodScopeFuncs` inside the top-level `BeforeEach`
block caused the tests using them in the `DescribeTable` to
fail with a nil pointer error.

The problem occurred because Ginkgo captures `Entry` parameters
when the Go package loads, whereas the `BeforeEach` block executes
at runtime. As a result, when the table Entries tried to use
`singleNUMAContainerScopeFuncs`, it was already nil because
the `BeforeEach` had not run yet to initialize it.

See also: https://github.com/openshift-kni/numaresources-operator/pull/2315

Signed-off-by: Francesco Romani <fromani@redhat.com>