e2e: move vars to avoid nil pointer dereference at runtime
This issue was a bit tricky to notice. The initialization of both ```singleNUMAContainerScopeFuncs``` and ```singleNUMAPodScopeFuncs``` inside the top-level BeforeEach block caused the tests using them in the ```DescribeTable``` to fail with a nil pointer error.

The problem occurred because Ginkgo captures ```Entry``` parameters when the Go package loads, whereas the BeforeEach block executes at runtime. As a result, when the table Entries tried to use ```singleNUMAContainerScopeFuncs```, it was already nil because the ```BeforeEach``` had not run yet to initialize it.

To fix it, we can simply move the initialization to the package level. This ensures the variables are initialized before Ginkgo parses the test structure.

Signed-off-by: Roy Shemtov <rshemtov@redhat.com>