CNF-20022: status: nrosched: set ObservedGeneration in all status conditions
The NUMAResourcesScheduler object status condition reports ObservedGeneration
only in the `DedicatedInformerActive` condition.
Let's expose it in all the relevant conditions, doing some minimal
internal refactoring along the way.
Note that we can't test this change using controller tests due both the
unique requirements of the operator and the way we (mis)use the
controller tests; we should add checks in (existing?) e2e tests, but we
postpone this to a later change.

Signed-off-by: Francesco Romani <fromani@redhat.com>