CNF-20021: status: nrooper: set ObservedGeneration in all status conditions
The NUMAResourcesScheduler object status condition never reports
ObservedGeneration in any status conditions, unlike the scheduler
object.

Likewise the scheduler change, let's expose it in all the relevant
conditions, doing some minimal internal refactoring along the way.

Note that we can't test this change using controller tests due both the
unique requirements of the operator and the way we (mis)use the
controller tests; we should add checks in (existing?) e2e tests, but we
postpone this to a later change.

Signed-off-by: Francesco Romani <fromani@redhat.com>