pkg: version: store discovered cluster full version
So far the application has been relying on the minimized cluster
version. While this is true for most general cases, delicate flows that
requires knowledge of the specific builds needs access to the full
version. Such flows is the one to enable the `Shared` mode of `SchedulerInformer` by
default starting specific fixed OCP builds.
This commit updates the discovery code to store the full version and
namely rearranges main.go to pass on the full version to internal reconciliation.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>