status: SS-conditions: update base conditions properly
The old version of the function was updating the condition inplace
without considering whether the it's a base condition or not.
The difference in base and non-base condition is that the status of one
base condition affect all others, thus the rest would need an update too.

The new version enhances this and splits the conditions into base and
non-base ones and update all conditions accordingly using u/s SetStatusCondition,
 which also cares to avoid noisy updates.
ref:
https://github.com/kubernetes/apimachinery/blob/master/pkg/api/meta/conditions.go

note that this peice is part of a larger enhancement that will soon
be used in the NRO-controller to update the conditions there.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>