test: remove GetOwnedBy
Getting deployments/daemonsets by owner is useful only in a handful
places, e.g. mostly (only) when checking the reliability of the
status reporting we added long ago.

While arguably using the OwnerReference is a even superior choice,
we made the decision long ago to directly report the dependant objects
in the status; so let's use that data instead of re-fetching it again.

Signed-off-by: Francesco Romani <fromani@redhat.com>