workflow: support Gemini code reviewer
Integrate gemini code reviewer to assist in initial review of the
PR. This is NOT meant to replace human review, but rather a way to
discover bugs earlier and provide potential performance improvements.
That said, project reviewers and PR submitters should review the
suggestions carefully and apply their best engineering judgment
considering the guidelines of the project.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>