e2e:mg: use image reference variable
introducing `E2E_NROP_MUSTGATHER_IMAGE_REF` env variable
which represents image name + tag/digest
and takes precedence over \`E2E_NROP_MUSTGATHER_IMAGE\` and \`E2E_NROP_MUSTGATHER_TAG\`.

The rationale for this change is to be able to provide
an image referrence using digest or without having the need to devide
it to image and tag (as prow CI does)
see: https://docs.ci.openshift.org/docs/architecture/ci-operator/#referring-to-images-in-ci-operator-configuration

however, we should keep the old behavior for backward compatibility and
we might consider obsolete that completely in the future.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>