unit-test: status: updating one base condition doesn't affect others
A base condition is one of the conditions that describe the status of
the CR: available, upgradeable, progressing, degraded.
a value of one is expected to update the other values as well, for
example we it doesn't make sense that a set of CR conditions has both
statuses of available and degraded as True.
This commit adds unit test coverage for these cases and fixes the tests values to reasonable ones.

Note that the new tests fail and this highlights the need to fix the
logic before relying further more on this function.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>