sched: manifest: disable dynamic resources plugin
The latest updates on kube scheduler enables Dynamic Resources plugin by
default. This is causing issue because the scheduler cluster role do not
have permission to list and watch ResourceClaims. The secondary
scheduler should not support DRA till further notice.
In this commit we chose to disable the plugin; If we want that to be
enabled we should remove the multi-point disabling and add RBAC
permissions to the NRT cluster role.

The main root cause for the failure is that the informer (responsible to
keep the cluster resources cache in sync) need to watch and list another
set of resources needed by the DynamicResources plugin. Having no RBAC
permissions make listing these resources fail and thus blocks the
informer flow, hence the scheduler doesn't continue to schedule pods.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>