serial: enhance TMScope tests to support compact cluster resources
Some set of the serial suite expects to have the resources set on the node in a
specific way ahead of running the suite. The tests are various and are configured
requesting specific amount of resources to be available on the cluster
node before running the test worloads. The values of the amounts are
hardcoded and that is by design. The problem with this is mainly the
baseload (consumed node resources that do not relate to the test).

On compact clusters the baseload is normally greater than on regular
MNO. And when the test runs with old hardcoded resources values, it
considers a greater baseload than usually is on regular MNO (initiialy
tests were built depending on this kind of clusters), and that affects
the amount of free resources that should be left free on the node(s),
causing to shake the "working" hardcoded values. This results in tests
failing due to unexpected behavior on compact clusters.

In this commit we consolidate tests that run on diffirent TM scopes but
they both expect same results (single container does not make a
difference).

Also we adjust the hardcoded values to ensure high enough gap to keep
the test stable. Having a ratios of current node resources is  problamatic
because it needs to be rounded and doesn't eventually turn out to fit into
the node and in many cases messes up the intented scenario.

The best we can do while still have full control over the requested
resources and the padding resources is tweak the resources amounts to
the minimum but stable as possible so that it can fit to run on compact
clusters. We prefer to have full control with logs over having
automatically computed resources especially when the cases and resources
are very diverse, which is good and makes a better coverage and that we
want to keep.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>